+++ src/lib/scriptengine/mod.rs
@@ -72,7 +72,7 @@ pub(crate) fn get_engine_type(task: &Task) -> EngineType {
 
 pub(crate) fn invoke(task: &Task, cli_arguments: &Vec<String>) -> bool {
     let engine_type = get_engine_type(&task);
-    let validate = !task.is_force();
+    let validate = !task.should_ignore_errors();
 
     match engine_type {
         EngineType::OS => {