+++ src/lib/types_test.rs
@@ -376,6 +376,7 @@ fn task_new() {
     assert!(task.description.is_none());
     assert!(task.category.is_none());
     assert!(task.workspace.is_none());
+    assert!(task.ignore_errors.is_none());
     assert!(task.force.is_none());
     assert!(task.env.is_none());
     assert!(task.cwd.is_none());
@@ -407,23 +408,45 @@ fn external_config_new() {
 }
 
 #[test]
-fn task_is_force_none() {
+fn task_should_ignore_errors_none() {
     let task = Task::new();
-    assert!(!task.is_force());
+    assert!(!task.should_ignore_errors());
 }
 
 #[test]
-fn task_is_force_false() {
+fn task_should_ignore_errors_false() {
+    let mut task = Task::new();
+    task.ignore_errors = Some(false);
+    assert!(!task.should_ignore_errors());
+}
+
+#[test]
+fn task_should_ignore_errors_true() {
+    let mut task = Task::new();
+    task.ignore_errors = Some(true);
+    assert!(task.should_ignore_errors());
+}
+
+#[test]
+fn task_should_ignore_errors_force_false() {
     let mut task = Task::new();
     task.force = Some(false);
-    assert!(!task.is_force());
+    assert!(!task.should_ignore_errors());
 }
 
 #[test]
-fn task_is_force_true() {
+fn task_should_ignore_errors_force_true() {
     let mut task = Task::new();
     task.force = Some(true);
-    assert!(task.is_force());
+    assert!(task.should_ignore_errors());
+}
+
+#[test]
+fn task_should_ignore_errors_false_force_true() {
+    let mut task = Task::new();
+    task.ignore_errors = Some(false);
+    task.force = Some(true);
+    assert!(!task.should_ignore_errors());
 }
 
 #[test]
@@ -433,7 +456,7 @@ fn task_extend_both_have_misc_data() {
     base.command = Some("test1".to_string());
     base.disabled = Some(false);
     base.private = Some(false);
-    base.watch = Some(false);
+    base.watch = Some(TaskWatchOptions::Boolean(false));
     base.script = Some(vec!["1".to_string(), "2".to_string()]);
 
     let extended = Task {
@@ -445,9 +468,10 @@ fn task_extend_both_have_misc_data() {
         workspace: None,
         disabled: Some(true),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
+        ignore_errors: Some(true),
         force: Some(true),
         env: Some(IndexMap::new()),
         cwd: None,
@@ -482,6 +506,7 @@ fn task_extend_both_have_misc_data() {
     assert!(base.watch.is_some());
     assert!(base.condition.is_none());
     assert!(base.condition_script.is_none());
+    assert!(base.ignore_errors.is_some());
     assert!(base.force.is_some());
     assert!(base.env.is_some());
     assert!(base.cwd.is_none());
@@ -509,7 +534,8 @@ fn task_extend_both_have_misc_data() {
     assert_eq!(base.command.unwrap(), "test1");
     assert!(base.disabled.unwrap());
     assert!(base.private.unwrap());
-    assert!(base.watch.unwrap());
+    assert_eq!(base.watch.unwrap(), TaskWatchOptions::Boolean(true));
+    assert!(base.ignore_errors.unwrap());
     assert!(base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 0);
     assert_eq!(base.alias.unwrap(), "alias2");
@@ -527,9 +553,10 @@ fn task_extend_extended_have_all_fields() {
         workspace: None,
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
+        ignore_errors: Some(true),
         force: Some(true),
         env: Some(IndexMap::new()),
         cwd: None,
@@ -563,7 +590,7 @@ fn task_extend_extended_have_all_fields() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
-        watch: Some(false),
+        watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -574,6 +601,7 @@ fn task_extend_extended_have_all_fields() {
             rust_version: None,
         }),
         condition_script: Some(vec!["exit 0".to_string()]),
+        ignore_errors: Some(false),
         force: Some(false),
         env: Some(env.clone()),
         cwd: Some("cwd".to_string()),
@@ -596,7 +624,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -607,6 +635,7 @@ fn task_extend_extended_have_all_fields() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -626,7 +655,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -637,6 +666,7 @@ fn task_extend_extended_have_all_fields() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -656,7 +686,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -667,6 +697,7 @@ fn task_extend_extended_have_all_fields() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -695,6 +726,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.watch.is_some());
     assert!(base.condition.is_some());
     assert!(base.condition_script.is_some());
+    assert!(base.ignore_errors.is_some());
     assert!(base.force.is_some());
     assert!(base.env.is_some());
     assert!(base.cwd.is_some());
@@ -725,8 +757,9 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.workspace.unwrap());
     assert!(base.disabled.unwrap());
     assert!(!base.private.unwrap());
-    assert!(!base.watch.unwrap());
+    assert_eq!(base.watch.unwrap(), TaskWatchOptions::Boolean(false));
     assert_eq!(base.condition_script.unwrap().len(), 1);
+    assert!(!base.ignore_errors.unwrap());
     assert!(!base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 1);
     assert_eq!(base.cwd.unwrap(), "cwd".to_string());
@@ -768,7 +801,7 @@ fn task_extend_clear_with_no_data() {
         workspace: Some(false),
         disabled: Some(true),
         private: Some(false),
-        watch: Some(false),
+        watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -779,6 +812,7 @@ fn task_extend_clear_with_no_data() {
             rust_version: None,
         }),
         condition_script: Some(vec!["exit 0".to_string()]),
+        ignore_errors: Some(false),
         force: Some(false),
         env: Some(env.clone()),
         cwd: Some("cwd".to_string()),
@@ -801,7 +835,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -812,6 +846,7 @@ fn task_extend_clear_with_no_data() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -831,7 +866,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -842,6 +877,7 @@ fn task_extend_clear_with_no_data() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -861,7 +897,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -872,6 +908,7 @@ fn task_extend_clear_with_no_data() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -902,6 +939,7 @@ fn task_extend_clear_with_no_data() {
     assert!(base.watch.is_none());
     assert!(base.condition.is_none());
     assert!(base.condition_script.is_none());
+    assert!(base.ignore_errors.is_none());
     assert!(base.force.is_none());
     assert!(base.env.is_none());
     assert!(base.cwd.is_none());
@@ -938,7 +976,7 @@ fn task_extend_clear_with_all_data() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
-        watch: Some(false),
+        watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -949,6 +987,7 @@ fn task_extend_clear_with_all_data() {
             rust_version: None,
         }),
         condition_script: Some(vec!["exit 0".to_string()]),
+        ignore_errors: Some(false),
         force: Some(false),
         env: Some(env.clone()),
         cwd: Some("cwd".to_string()),
@@ -971,7 +1010,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -982,6 +1021,7 @@ fn task_extend_clear_with_all_data() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -1001,7 +1041,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1012,6 +1052,7 @@ fn task_extend_clear_with_all_data() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -1031,7 +1072,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1042,6 +1083,7 @@ fn task_extend_clear_with_all_data() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env.clone()),
             cwd: Some("cwd".to_string()),
@@ -1069,6 +1111,7 @@ fn task_extend_clear_with_all_data() {
     assert!(base.watch.is_some());
     assert!(base.condition.is_some());
     assert!(base.condition_script.is_some());
+    assert!(base.ignore_errors.is_some());
     assert!(base.force.is_some());
     assert!(base.env.is_some());
     assert!(base.cwd.is_some());
@@ -1138,9 +1181,10 @@ fn task_get_normalized_task_undefined() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
+        ignore_errors: None,
         force: None,
         env: None,
         cwd: None,
@@ -1171,6 +1215,7 @@ fn task_get_normalized_task_undefined() {
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());
+    assert!(normalized_task.ignore_errors.is_none());
     assert!(normalized_task.force.is_none());
     assert!(normalized_task.env.is_none());
     assert!(normalized_task.cwd.is_none());
@@ -1203,7 +1248,11 @@ fn task_get_normalized_task_undefined() {
     assert!(!normalized_task.workspace.unwrap());
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
-    assert!(normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(true)
+    );
+    assert!(!normalized_task.ignore_errors.unwrap_or(false));
     assert!(!normalized_task.force.unwrap_or(false));
     assert_eq!(normalized_task.alias.unwrap(), "alias");
     assert_eq!(normalized_task.linux_alias.unwrap(), "linux");
@@ -1243,7 +1292,7 @@ fn task_get_normalized_task_with_override_no_clear() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1254,6 +1303,7 @@ fn task_get_normalized_task_with_override_no_clear() {
             rust_version: None,
         }),
         condition_script: Some(vec!["exit 0".to_string()]),
+        ignore_errors: Some(false),
         force: Some(false),
         env: Some(IndexMap::new()),
         cwd: Some("cwd".to_string()),
@@ -1272,7 +1322,7 @@ fn task_get_normalized_task_with_override_no_clear() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1283,6 +1333,7 @@ fn task_get_normalized_task_with_override_no_clear() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["exit 0".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env),
             cwd: Some("cwd2".to_string()),
@@ -1318,6 +1369,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
+    assert!(normalized_task.ignore_errors.is_some());
     assert!(normalized_task.force.is_some());
     assert!(normalized_task.env.is_some());
     assert!(normalized_task.cwd.is_some());
@@ -1348,8 +1400,12 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
-    assert!(!normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(false)
+    );
     assert_eq!(normalized_task.condition_script.unwrap().len(), 1);
+    assert!(normalized_task.ignore_errors.unwrap());
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
     assert_eq!(normalized_task.cwd.unwrap(), "cwd2".to_string());
@@ -1391,7 +1447,7 @@ fn task_get_normalized_task_with_override_clear_false() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1402,6 +1458,7 @@ fn task_get_normalized_task_with_override_clear_false() {
             rust_version: None,
         }),
         condition_script: Some(vec!["exit 0".to_string()]),
+        ignore_errors: Some(false),
         force: Some(false),
         env: Some(IndexMap::new()),
         cwd: Some("cwd".to_string()),
@@ -1419,7 +1476,7 @@ fn task_get_normalized_task_with_override_clear_false() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
-            watch: Some(false),
+            watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
                 platforms: Some(vec!["linux".to_string()]),
@@ -1434,6 +1491,7 @@ fn task_get_normalized_task_with_override_clear_false() {
                 rust_version: None,
             }),
             condition_script: Some(vec!["echo test".to_string(), "exit 1".to_string()]),
+            ignore_errors: Some(true),
             force: Some(true),
             env: Some(env),
             cwd: Some("cwd2".to_string()),
@@ -1469,6 +1527,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
+    assert!(normalized_task.ignore_errors.is_some());
     assert!(normalized_task.force.is_some());
     assert!(normalized_task.env.is_some());
     assert!(normalized_task.cwd.is_some());
@@ -1499,8 +1558,12 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
-    assert!(!normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(false)
+    );
     assert_eq!(normalized_task.condition_script.unwrap().len(), 2);
+    assert!(normalized_task.ignore_errors.unwrap());
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
     assert_eq!(normalized_task.cwd.unwrap(), "cwd2".to_string());
@@ -1537,7 +1600,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1548,6 +1611,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
             rust_version: None,
         }),
         condition_script: Some(vec!["exit 0".to_string()]),
+        ignore_errors: Some(false),
         force: Some(false),
         env: Some(IndexMap::new()),
         cwd: Some("cwd".to_string()),
@@ -1572,6 +1636,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
             watch: None,
             condition: None,
             condition_script: None,
+            ignore_errors: None,
             force: None,
             env: None,
             cwd: None,
@@ -1599,6 +1664,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
+    assert!(normalized_task.ignore_errors.is_some());
     assert!(normalized_task.force.is_some());
     assert!(normalized_task.env.is_some());
     assert!(normalized_task.cwd.is_some());
@@ -1628,7 +1694,11 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert_eq!(normalized_task.command.unwrap(), "command");
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
-    assert!(normalized_task.watch.unwrap());
+    assert_eq!(
+        normalized_task.watch.unwrap(),
+        TaskWatchOptions::Boolean(true)
+    );
+    assert!(!normalized_task.ignore_errors.unwrap());
     assert!(!normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 0);
     assert_eq!(normalized_task.cwd.unwrap(), "cwd".to_string());
@@ -1661,7 +1731,7 @@ fn task_get_normalized_task_with_override_clear_true() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
-        watch: Some(true),
+        watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
@@ -1672,6 +1742,7 @@ fn task_get_normalized_task_with_override_clear_true() {
             rust_version: None,
         }),
         condition_script: Some(vec!["exit 0".to_string()]),
+        ignore_errors: Some(false),
         force: Some(false),
         env: Some(IndexMap::new()),
         cwd: Some("cwd".to_string()),
@@ -1696,6 +1767,7 @@ fn task_get_normalized_task_with_override_clear_true() {
             watch: None,
             condition: None,
             condition_script: None,
+            ignore_errors: None,
             force: None,
             env: None,
             cwd: None,
@@ -1723,6 +1795,7 @@ fn task_get_normalized_task_with_override_clear_true() {
     assert!(normalized_task.watch.is_none());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());
+    assert!(normalized_task.ignore_errors.is_none());
     assert!(normalized_task.force.is_none());
     assert!(normalized_task.env.is_none());
     assert!(normalized_task.cwd.is_none());